requirements.txt
streamlit==1.32.2
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-experimental>=0.0.52
tabulate>=0.9.0
tabula-py>=2.9.0
streamlit-authenticator>=0.2.3
PyYAML>=6.0
